Move AnnotatedString.transform implementation to jsMain source set
