+++ compose/ui/ui-text/src/jsNativeMain/kotlin/androidx/compose/ui/text/AnnotatedString.native.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,8 +19,6 @@ package androidx.compose.ui.text
 import androidx.compose.ui.text.AnnotatedString.Range
 import androidx.compose.ui.util.fastMap
 
-// import java.util.SortedSet
-
 /**
  * The core function of [AnnotatedString] transformation.
  *
@@ -29,37 +27,37 @@ import androidx.compose.ui.util.fastMap
  */
 internal actual fun AnnotatedString.transform(
     transform: (String, Int, Int) -> String
-): AnnotatedString = TODO("implement native AnnotatedString.transform")
+): AnnotatedString {
 
-
-/*
-{
-    val transitions = sortedSetOf(0, text.length)
-    collectRangeTransitions(spanStyles, transitions)
-    collectRangeTransitions(paragraphStyles, transitions)
+    // Sorted set is unavailable in Kotlin Native standard library, so sorting regular list instead.
+    val transitions = mutableListOf(0, text.length)
+    collectRangeTransitions(spanStylesOrNull, transitions)
+    collectRangeTransitions(paragraphStylesOrNull, transitions)
+    collectRangeTransitions(annotations, transitions)
 
     var resultStr = ""
     val offsetMap = mutableMapOf(0 to 0)
+    transitions.sort()
     transitions.windowed(size = 2) { (start, end) ->
         resultStr += transform(text, start, end)
         offsetMap.put(end, resultStr.length)
     }
 
-    val newSpanStyles = spanStyles.fastMap {
+    val newSpanStyles = spanStylesOrNull?.fastMap {
         // The offset map must have mapping entry from all style start, end position.
         Range(it.item, offsetMap[it.start]!!, offsetMap[it.end]!!)
     }
-    val newParaStyles = paragraphStyles.fastMap {
+    val newParaStyles = paragraphStylesOrNull?.fastMap {
         Range(it.item, offsetMap[it.start]!!, offsetMap[it.end]!!)
     }
-    val newAnnotations = annotations.fastMap {
+    val newAnnotations = annotations?.fastMap {
         Range(it.item, offsetMap[it.start]!!, offsetMap[it.end]!!)
     }
 
     return AnnotatedString(
         text = resultStr,
-        spanStyles = newSpanStyles,
-        paragraphStyles = newParaStyles,
+        spanStylesOrNull = newSpanStyles,
+        paragraphStylesOrNull = newParaStyles,
         annotations = newAnnotations
     )
 }
@@ -70,15 +68,14 @@ internal actual fun AnnotatedString.transform(
  * @param ranges The list of AnnotatedString.Range
  * @param target The output list
  */
-private fun <T> collectRangeTransitions(
-    ranges: List<Range<T>>,
-    target: SortedSet<Int>
+private fun collectRangeTransitions(
+    ranges: List<Range<*>>?,
+    target: MutableCollection<Int>
 ) {
-    ranges.fastFold(target) { acc, range ->
+    ranges?.fastFold(target) { acc, range ->
         acc.apply {
             add(range.start)
             add(range.end)
         }
     }
 }
-*/
